+++ unity-package/Runtime/Data/V1/CustomAccessory_V1.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace MoreCustomizations.Data {
@@ -11,11 +12,19 @@ namespace MoreCustomizations.Data {
         
         [field: SerializeField]
         public Texture Texture { get; internal set; }
-
+        
         public override Texture IconTexture
             => Texture;
-
-        public override bool IsValid()
-            => Texture;
+        
+        public override IEnumerable<ValidateStatus> GetValidateContentStatuses() {
+            
+            if (!Texture) {
+                
+                yield return ValidateStatus.Error($"{nameof(Texture)} is empty.");
+                yield break;
+            }
+            
+            yield return ValidateStatus.Valid;
+        }
     }
 }