+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change logs
 
+## 1.1.4
+
+### Breaking change
+
+Tool package version is now up to `1.1.4` due fit support and UX improvement for tool package.  
+Please update it on Package Manager window in Unity.
+
+- Added detailed validation for tool package.
+- Added fit support. [#12](https://github.com/Creta5164/peak-more-customizations/pull/12), [#14](https://github.com/Creta5164/peak-more-customizations/pull/14)  
+  Contributed by Umbrason
+
 ## 1.1.3
 
 - Fixed name conflict in contents by adding hash from bundle path.